#245: Fixed PlanDefinition bundling still requiring source directory arguments
